+++ hvega/CHANGELOG.md
@@ -1,6 +1,12 @@
 For the latest version of this document, please see
 [https://github.com/DougBurke/hvega/blob/master/hvega/CHANGELOG.md](https://github.com/DougBurke/hvega/blob/master/hvega/CHANGELOG.md).
 
+## 0.2
+
+Rename Left and Right constructors to avoid name clashes (breaking API change).
+
+Clarify how to use the library in the main `hvega` module.
+
 ## 0.1.0.3
 
 The only change is to the cabal file, where `cabal-version: >=1.18` has